fix(purpuse): assets path
