+++ client/src/App.tsx
@@ -5,7 +5,7 @@ import { GlobalStyle } from './styles/global';
 
 
 import { Home } from './pages/home';  
-import { Purpose } from './pages/purpose';
+import { Purpose } from './components/purpose';
 
 function App() {
   return (
